+++ omniisaacgymenvs/utils/config_utils/sim_config.py
@@ -27,6 +27,7 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
+from omni.isaac.core.utils.extensions import enable_extension
 from omniisaacgymenvs.utils.config_utils.default_scene_params import *
 import copy
 import omni.usd
@@ -46,15 +47,14 @@ class SimConfig():
         if self._config["test"] == True:
             self._sim_params["enable_scene_query_support"] = True
 
-        from omni.isaac.core.utils.extensions import enable_extension
         if self._config["headless"] == True and not self._sim_params["enable_cameras"] and not self._config["enable_livestream"]:
             self._sim_params["use_fabric"] = False
             self._sim_params["enable_viewport"] = False
         else:
             self._sim_params["enable_viewport"] = True
             enable_extension("omni.kit.viewport.bundle")
-        enable_extension("omni.replicator.isaac")
 
+        self._sim_params["warp"] = self._config["warp"]
         if self._sim_params["disable_contact_processing"]:
             carb.settings.get_settings().set_bool("/physics/disableContactProcessing", True)