+++ omniisaacgymenvs/utils/task_util.py
@@ -33,42 +33,53 @@ def initialize_task(config, env, init_sim=True):
 
     from omniisaacgymenvs.tasks.allegro_hand import AllegroHandTask
     from omniisaacgymenvs.tasks.ant import AntLocomotionTask
+    from omniisaacgymenvs.tasks.warp.ant import AntLocomotionTask as AntLocomotionTaskWarp
     from omniisaacgymenvs.tasks.anymal import AnymalTask
     from omniisaacgymenvs.tasks.anymal_terrain import AnymalTerrainTask
     from omniisaacgymenvs.tasks.ball_balance import BallBalanceTask
     from omniisaacgymenvs.tasks.cartpole import CartpoleTask
+    from omniisaacgymenvs.tasks.warp.cartpole import CartpoleTask as CartpoleTaskWarp
     from omniisaacgymenvs.tasks.factory.factory_task_nut_bolt_pick import FactoryTaskNutBoltPick
     from omniisaacgymenvs.tasks.franka_cabinet import FrankaCabinetTask
     from omniisaacgymenvs.tasks.humanoid import HumanoidLocomotionTask
+    from omniisaacgymenvs.tasks.warp.humanoid import HumanoidLocomotionTask as HumanoidLocomotionTaskWarp
     from omniisaacgymenvs.tasks.ingenuity import IngenuityTask
     from omniisaacgymenvs.tasks.quadcopter import QuadcopterTask
     from omniisaacgymenvs.tasks.shadow_hand import ShadowHandTask
     from omniisaacgymenvs.tasks.crazyflie import CrazyflieTask
-    
-    # Mappings from strings to environments
-    task_map = {
-        "AllegroHand": AllegroHandTask,
-        "Ant": AntLocomotionTask,
-        "Anymal": AnymalTask,
-        "AnymalTerrain": AnymalTerrainTask,
-        "BallBalance": BallBalanceTask,
-        "Cartpole": CartpoleTask,
-        "FactoryTaskNutBoltPick": FactoryTaskNutBoltPick,
-        "FrankaCabinet": FrankaCabinetTask,
-        "Humanoid": HumanoidLocomotionTask,
-        "Ingenuity": IngenuityTask,
-        "Quadcopter": QuadcopterTask,
-        "Crazyflie": CrazyflieTask,
-        "ShadowHand": ShadowHandTask,
-        "ShadowHandOpenAI_FF": ShadowHandTask,
-        "ShadowHandOpenAI_LSTM": ShadowHandTask,
-    }
 
     cfg = sim_config.config
+    if cfg["warp"]:
+        task_map = {
+            "Cartpole": CartpoleTaskWarp,
+            "Ant":AntLocomotionTaskWarp,
+            "Humanoid": HumanoidLocomotionTaskWarp
+        }
+    else:
+        # Mappings from strings to environments
+        task_map = {
+            "AllegroHand": AllegroHandTask,
+            "Ant": AntLocomotionTask,
+            "Anymal": AnymalTask,
+            "AnymalTerrain": AnymalTerrainTask,
+            "BallBalance": BallBalanceTask,
+            "Cartpole": CartpoleTask,
+            "FrankaCabinet": FrankaCabinetTask,
+            "FactoryTaskNutBoltPick": FactoryTaskNutBoltPick,
+            "Humanoid": HumanoidLocomotionTask,
+            "Ingenuity": IngenuityTask,
+            "Quadcopter": QuadcopterTask,
+            "Crazyflie": CrazyflieTask,
+            "ShadowHand": ShadowHandTask,
+            "ShadowHandOpenAI_FF": ShadowHandTask,
+            "ShadowHandOpenAI_LSTM": ShadowHandTask,
+        }
+
     task = task_map[cfg["task_name"]](
         name=cfg["task_name"], sim_config=sim_config, env=env
     )
 
-    env.set_task(task=task, sim_params=sim_config.get_physics_params(), backend="torch", init_sim=init_sim)
+    backend = "warp" if cfg["warp"] else "torch"
+    env.set_task(task=task, sim_params=sim_config.get_physics_params(), backend=backend, init_sim=init_sim)
 
     return task